+++ core/common/src/commonMain/kotlin/com/chimali/core/common/di/DataStoreModule.kt
@@ -1,0 +1,5 @@
+package com.chimali.core.common.di
+
+import org.koin.core.module.Module
+
+expect val dataStoreModule: Module